+++ src/ui/tab_bar.rs
@@ -193,6 +193,7 @@ impl TabBar {
   pub fn new(
     id: Id,
     cap: &mut dyn MutCap<Event, Message>,
+    detail_dialog: Id,
     tag_dialog: Id,
     in_out: Id,
     tasks: Rc<Tasks>,
@@ -216,7 +217,15 @@ impl TabBar {
           id,
           Box::new(|| Box::new(TaskListBoxData::new(tasks, view, toggle_tag))),
           Box::new(move |id, cap| {
-            Box::new(TaskListBox::new(id, cap, tab_bar, tag_dialog, in_out, task))
+            Box::new(TaskListBox::new(
+              id,
+              cap,
+              tab_bar,
+              detail_dialog,
+              tag_dialog,
+              in_out,
+              task,
+            ))
           }),
         );