+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added support for editing of multi-line free-form task "details"
 - Fixed potential panic caused by clipping task summaries at non-char
   boundaries
 - Bumped `icalendar` dependency to `0.16`